Introduce Level 2 hot non-profiled nmethod segment
Based on the existing NonProfiledHotCodeHeap segement, introduce a
similar L2HotNonProfiled segment at a lower level.

Similar to the former, use the AllocInNonProfiledHotCodeHeap attribute
in the CompilerDirectivesFile to control the JIT in placing certain
nmethod into these two segements. Introduce "HotLevel" attribute to
determine which one to place them in (1: NonProfiledHotCode,
2: L2HotNonProfiled).